# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: TeamFlow
  version: '1.0'
servers:
  -
    url: 'http://teamflow-be.gplans.it'
    description: 'API Server'
paths:
...
  /api/todo/all:
    get:
      tags:
        - todo
      summary: 'Get all todos data'
      operationId: getAllTodos
      parameters:
        -
          name: limit
          in: query
          description: 'Limit of elements'
          required: false
          schema:
            type: integer
            example: 10
        -
          name: page
          in: query
          description: 'Current page'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: sortBy
          in: query
          description: 'sort by element'
          required: false
          schema:
            type: string
            example: created_at
        -
          name: sortValue
          in: query
          description: 'sorting type'
          required: false
          schema:
            type: string
            example: descend
        -
          name: filterBy
          in: query
          description: 'fields to filter'
          required: false
          schema:
            type: string
            example: ''
        -
          name: filterValue
          in: query
          description: 'Values to filter'
          required: false
          schema:
            type: string
            example: ''
        -
          name: start
          in: query
          description: 'filter by date: start data'
          required: false
          schema:
            type: string
            example: ''
        -
          name: end
          in: query
          description: 'filter by date: end data'
          required: false
          schema:
            type: string
            example: ''
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TodosResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
      security:
        -
          bearerAuth: []
  /api/todo/shared/all:
    get:
      tags:
        - todo
      summary: 'Get all shared todos'
      operationId: getAllSharedTodos
      parameters:
        -
          name: limit
          in: query
          description: 'Limit of elements'
          required: false
          schema:
            type: integer
            example: 10
        -
          name: page
          in: query
          description: 'Current page'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: sortBy
          in: query
          description: 'sort by element'
          required: false
          schema:
            type: string
            example: created_at
        -
          name: sortValue
          in: query
          description: 'sorting type'
          required: false
          schema:
            type: string
            example: descend
        -
          name: filterBy
          in: query
          description: 'fields to filter'
          required: false
          schema:
            type: string
            example: ''
        -
          name: filterValue
          in: query
          description: 'Values to filter'
          required: false
          schema:
            type: string
            example: ''
        -
          name: start
          in: query
          description: 'filter by date: start data'
          required: false
          schema:
            type: string
            example: ''
        -
          name: end
          in: query
          description: 'filter by date: end data'
          required: false
          schema:
            type: string
            example: ''
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TodosResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
      security:
        -
          bearerAuth: []
  '/api/todo/{todoId}/accounts/all':
    get:
      tags:
        - todo
      summary: 'Get all todo accounts'
      description: 'Get all accounts of a specific todo'
      operationId: getAllTodoAccounts
      parameters:
        -
          name: todoId
          in: path
          description: 'todo id'
          required: true
          schema:
            type: string
            example: 2588e5ad-0c5c-44e7-a865-a196b067c621
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodoAccountsResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Todo not found'
                type: object
      security:
        -
          bearerAuth: []
...
  /api/todo:
    post:
      tags:
        - todo
      summary: 'Create new todo'
      description: ''
      operationId: createTodo
      requestBody:
        description: 'JSON with todo data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoBodyReq'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: string
                        example: 45ca2e2c-91d1-482c-85bf-f0fc23022808
                    type: object
                  message:
                    type: string
                    example: 'Insert success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Insert unsuccess'
                type: object
      security:
        -
          bearerAuth: []
components:
  schemas:
...
    CommentResponse:
      properties:
        id:
          type: string
          example: 6ca498b0-3fb6-4094-ba16-42728383c587
        content:
          type: string
          example: 'Commento numero uno'
        created_at:
          type: string
          example: '2025-01-18 16:14:44'
        updated_at:
          type: string
          example: '2025-01-18 16:14:45'
        todo_id:
          type: string
          example: 1dcc7661-1251-456a-873c-2aa2028de9fd
        account_id:
          type: string
          example: 8a587029-80a2-4ae9-82e6-4f69f7383e63
        account_username:
          type: string
          example: test
      type: object
...
    TodoResponse:
      properties:
        id:
          type: string
          example: 1dcc7661-1251-456a-873c-2aa2028de9fd
        title:
          type: string
          example: 'Andare in palestra'
        description:
          type: string
          example: 'Ricorda la cintura'
        note:
          type: string
        category:
          type: string
          example: hobby
        checked:
          type: boolean
        created_at:
          type: string
          example: '2024-10-26 15:40:00'
        updated_at:
          type: string
          example: '2024-10-26 15:40:00'
        account_id:
          type: string
          example: 8a587029-80a2-4ae9-82e6-4f69f7383e63
        sharedWith:
          type: array
          items:
            properties:
              id:
                type: string
                example: 2d414d58-79d9-4b90-897f-d29889e86b98
              username:
                type: string
                example: account_test
            type: object
        isShared:
          type: boolean
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
      type: object
    TodoAccountsResponse:
      properties:
        id:
          type: string
          example: 1dcc7661-1251-456a-873c-2aa2028de9fd
        username:
          type: string
          example: user_00
      type: object
    TodosResponse:
      properties:
        current_page:
          type: number
          example: 1
        data:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
        first_page_url:
          type: string
          example: 'http://localhost:8000/api/todo/all?page=1'
        from:
          type: number
          example: 1
        last_page:
          type: number
          example: 1
        last_page_url:
          type: string
          example: 'http://localhost:8000/api/todo/all?page=1'
        links:
          type: array
          items:
            properties:
              url:
                type: string
                format: nullable
              label:
                type: string
                example: pagination.previous
              active:
                type: boolean
            type: object
        next_page_url:
          type: string
          format: nullable
        path:
          type: string
          example: 'http://localhost:8000/api/todo/all'
        per_page:
          type: number
          example: 10
        prev_page_url:
          type: string
          format: nullable
        to:
          type: number
          example: 1
        total:
          type: number
          example: 1
      type: object
    TodoBodyReq:
      properties:
        title:
          type: string
          example: 'Todo di prova'
        description:
          type: string
          example: 'descrizione di prova'
        note:
          type: string
          example: 'nessuna nota'
        category:
          type: string
          example: importanti
        checked:
          type: boolean
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...